moved load_most_recent_games into dependencies.py
